+++ app/models/element.rb
@@ -2,6 +2,7 @@ class Element < ApplicationRecord
   include ScoreVirtualAttributes
   
   ## relations
+  #has_many :element_judge_details, dependent: :destroy
   belongs_to :score
 
   ## scopes