+++ app/models/score.rb
@@ -1,4 +1,6 @@
 class Score < ApplicationRecord
+  before_save :set_score_name
+  
   ## relations
   has_many :elements, dependent: :destroy, autosave: true
   has_many :components, dependent: :destroy, autosave: true
@@ -44,16 +46,6 @@ class Score < ApplicationRecord
   scope :segment, ->(s){ where(segment: s) }
 
   ##
-  def update(parsed)
-    attrs = self.class.column_names.map(&:to_sym) & parsed.keys
-    self.attributes = parsed.slice(*attrs)
-    set_score_name
-    ActiveRecord::Base.transaction {
-      save!
-      parsed[:elements].map {|e| elements.create(e)}
-      parsed[:components].map {|e| components.create(e)}
-    }
-  end
   def summary
     skater_name = self.skater.try(:name) || self.skater_name
     nation = self.skater.try(:nation) || self.nation
@@ -64,12 +56,12 @@ class Score < ApplicationRecord
 
   private
   def set_score_name
-    return if name.present?
+    if name.blank?
+      category_abbr = Category.find_by(name: category).try(:abbr)
+      segment_abbr = segment.to_s.split(/ +/).map {|d| d[0]}.join # e.g. 'SHORT PROGRAM' => 'SP'
 
-    category_abbr = Category.find_by(name: category).try(:abbr)
-    segment_abbr = segment.to_s.split(/ +/).map {|d| d[0]}.join # e.g. 'SHORT PROGRAM' => 'SP'
-
-    self.name = [competition.try(:short_name), category_abbr, segment_abbr, ranking].join('-')
+      self.name = [competition.try(:short_name), category_abbr, segment_abbr, ranking].join('-')
+    end
     self
   end
 end