+++ app/parsers/competition_parser/wtt2017.rb
@@ -1,6 +1,6 @@
-class Parsers
+module CompetitionParser
   class Wtt2017 < IsuGeneric
-    class CompetitionParser < IsuGeneric::CompetitionParser
+    class SummaryParser < IsuGeneric::SummaryParser
       def parse_name(_page)
         "ISU World Team Trophy 2017"
       end
@@ -43,7 +43,7 @@ class Parsers
         summary
       end
       # rubocop:disable all
-      def parse_time_schedule(_page)
+      def parse_time_schedule(_page, date_format:)
         Time.zone ||= "UTC"
         [
          {