+++ config/routes.rb
@@ -34,6 +34,8 @@ Rails.application.routes.draw do
     end
   end
 
+  resources :analysis  # for temtative panel analysis
+
   resources :statics, only: :index
 
   resources :scorecalc, only: :index do